Get rid of unnecessary CachedConverter interface
