+++ tests/Converter/CachedConverterTest.php
@@ -6,7 +6,7 @@ namespace Neusta\ConverterBundle\Tests\Converter;
 
 use Neusta\ConverterBundle\CacheManagement\DefaultCacheManagement;
 use Neusta\ConverterBundle\Converter\Converter;
-use Neusta\ConverterBundle\Converter\DefaultCachedConverter;
+use Neusta\ConverterBundle\Converter\CachedConverter;
 use Neusta\ConverterBundle\Converter\DefaultConverter;
 use Neusta\ConverterBundle\Tests\Fixtures\CacheManagement\UserKeyFactory;
 use Neusta\ConverterBundle\Tests\Fixtures\Factory\PersonFactory;
@@ -15,14 +15,14 @@ use Neusta\ConverterBundle\Tests\Fixtures\Model\User;
 use Neusta\ConverterBundle\Tests\Fixtures\Populator\PersonNamePopulator;
 use PHPUnit\Framework\TestCase;
 
-class DefaultCachedConverterTest extends TestCase
+class CachedConverterTest extends TestCase
 {
     /** @var Converter<User, Person> */
     private Converter $converter;
 
     protected function setUp(): void
     {
-        $this->converter = new DefaultCachedConverter(
+        $this->converter = new CachedConverter(
             new DefaultConverter(
                 new PersonFactory(),
                 [
